Jest encounter unexpected token: move getCategoryRouteName out of helper function
